refactor rename activatebody and permissionsbody
